// src/generated/model/academicYearResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AcademicYearResponseStatus } from './academicYearResponseStatus.ts';

export interface AcademicYearResponse {
  id?: string;
  name?: string;
  code?: string;
  description?: string;
  startDate?: string;
  endDate?: string;
  status?: AcademicYearResponseStatus;
  isCurrentYear?: boolean;
  createdAt?: string;
  updatedAt?: string;
}